fix: rename appServer.ts model file to app-server.ts
